backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
pydantic-settings>=2.0.0
